# libs/expense_tracker_api/src/pot_api.rs
pub mod pot_api {
    use crate::api::{check_error, get_sub_claim, ApiResponse};
    use crate::currency_api::currency_api::CurrencyDTO;
    use std::sync::Arc;

    use crate::expense_api::expense_api::{ExpenseDTO, NewExpenseDTO};
    use crate::user_api::user_api::UserDTO;
    use axum::extract::{Path, State};
    use axum::http::request::Parts;
    use axum::http::StatusCode;
    use axum::Json;
    use chrono::{DateTime, Utc};
    use expense_tracker_db::pots::pots::{NewPot, Pot, PotToUser};
    use expense_tracker_db::setup::DbPool;
    use expense_tracker_db::users::users::User;
    use expense_tracker_services::currency_service::currency_service;
    use expense_tracker_services::currency_service::currency_service::CurrencyService;
    use expense_tracker_services::expense_service::expense_service;
    use expense_tracker_services::expense_service::expense_service::ExpenseService;
    use expense_tracker_services::pot_service::pot_service;
    use expense_tracker_services::pot_service::pot_service::PotService;
    use serde::{Deserialize, Serialize};
    use utoipa::ToSchema;
    use utoipa_axum::router::OpenApiRouter;
    use utoipa_axum::routes;
    use uuid::Uuid;

    /// Holds the App State for the PotAPI.
    pub struct PotApiState {
        pot_service: PotService,
        currency_service: CurrencyService,
        expense_service: ExpenseService,
    }

    /// Registers all functions of the Pot API.
    pub fn register(pool: DbPool) -> OpenApiRouter {
        let shared_state = Arc::new(PotApiState {
            pot_service: pot_service::new_service(pool.clone()),
            currency_service: currency_service::new_service(pool.clone()),
            expense_service: expense_service::new_service(pool.clone()),
        });

        OpenApiRouter::new()
            .routes(routes!(create_pot))
            .routes(routes!(get_pots))
            .routes(routes!(add_expense))
            .routes(routes!(get_pot_expenses))
            .routes(routes!(add_users_to_pot))
            .routes(routes!(remove_user_from_pot))
            .routes(routes!(delete_pot))
            .routes(routes!(archive))
            .routes(routes!(unarchive))
            .routes(routes!(pay_pot))
            .with_state(shared_state)
    }

    /// DTO used when working with existing Pots.
    #[derive(ToSchema, Serialize)]
    pub struct PotDTO {
        id: i32,
        owner_id: Uuid,
        name: String,
        default_currency: CurrencyDTO,
        users: Vec<UserDTO>,
        /// Indicates the amount of money, the user is owed or owes others. If positive, other users
        /// need to pay that amount to the user. If negative, the user has to pay the given amount.
        net_balance: f64,
        archived: bool,
        created_at: DateTime<Utc>,
        archived_at: Option<DateTime<Utc>>
    }

    impl PotDTO {
        /// Creates a new PotDTO from a db Pot.
        pub fn from(
            pot: Pot,
            default_currency: CurrencyDTO,
            users: Vec<UserDTO>,
            net_balance: f64) -> Self {
            Self {
                id: pot.id(),
                owner_id: pot.owner_id(),
                name: pot.name().to_string(),
                default_currency,
                users,
                net_balance,
                archived: pot.is_archived(),
                created_at: pot.created_at(),
                archived_at: pot.archived_at()
            }
        }

        /// Create a vec<PotDTO> from a vec<Pot>.
        pub fn from_vec(
            pot_vec: Vec<(Pot, Vec<User>, f64)>,
            currency_vec: Vec<CurrencyDTO>,
        ) -> Vec<Self> {
            let mut dtos: Vec<PotDTO> = vec![];

            for pot in pot_vec {
                let pot_currency = currency_vec
                    .iter()
                    .find(|c| c.id() == pot.0.default_currency_id());

                if let Some(pot_currency) = pot_currency {
                    dtos.push(PotDTO::from(
                        pot.0,
                        (*pot_currency).clone(),
                        UserDTO::from_vec(pot.1),
                        pot.2
                    ))
                }
            }

            dtos
        }
    }

    /// DTO used when creating a new Pot.
    #[derive(ToSchema, Serialize, Deserialize)]
    pub struct NewPotDTO {
        #[schema(max_length=24)]
        name: String,
        default_currency_id: i32,
    }

    impl NewPotDTO {
        /// Converts the DTO to the db object.
        fn to_db(&self, owner_id: Uuid) -> NewPot {
            NewPot::new(owner_id, self.name.clone(), self.default_currency_id)
        }
    }

    #[derive(ToSchema, Serialize, Deserialize)]
    pub struct AddUserToPotDTO {
        user_id: Uuid,
    }

    impl AddUserToPotDTO {
        pub fn user_id(&self) -> Uuid {
            self.user_id
        }
    }

    #[derive(ToSchema, Serialize, Deserialize)]
    pub struct RemoveUserFromPotDTO {
        user_id: Uuid,
    }

    impl RemoveUserFromPotDTO {
        pub fn user_id(&self) -> Uuid {
            self.user_id
        }
    }

    /// Creates a pot from the given DTO for the bearer.
    #[utoipa::path(
        post,
        path = "/pots",
        tag = "Pots",
        responses(
            (status = 201, description = "The pot has been created", body = PotDTO),
        ),
        request_body = NewPotDTO,
        security(
            ("bearer" = [])
        )
    )]
    pub async fn create_pot(
        State(pot_api_state): State<Arc<PotApiState>>,
        parts: Parts,
        Json(new_pot): Json<NewPotDTO>,
    ) -> Result<ApiResponse<PotDTO>, ApiResponse<String>> {
        let subject_id = get_sub_claim(&parts)?;

        let result = pot_api_state
            .pot_service
            .create_pot(new_pot.to_db(subject_id))
            .await
            .map_err(check_error)?;

        Ok((
            StatusCode::CREATED,
            Json(PotDTO::from(
                result.0,
                CurrencyDTO::from(result.1),
                UserDTO::from_vec(result.2),
                0.0
            )),
        ))
    }

    /// Gets the list of all pots the bearer can view.
    #[utoipa::path(
        get,
        path = "/pots",
        tag = "Pots",
        responses(
            (status = 200, description = "The list of known pots.", body = Vec<PotDTO>)
        ),
        security(
                ("bearer" = [])
        )
    )]
    pub async fn get_pots(
        State(pot_api_state): State<Arc<PotApiState>>,
        parts: Parts,
    ) -> Result<ApiResponse<Vec<PotDTO>>, ApiResponse<String>> {
        let subject_id = get_sub_claim(&parts)?;

        let loaded_pots = pot_api_state
            .pot_service
            .get_pots(subject_id)
            .await
            .map_err(check_error)?;

        let all_currencies = pot_api_state
            .currency_service
            .get_currencies()
            .await
            .map_err(check_error)?;

        let all_currencies = CurrencyDTO::from_vec(all_currencies);

        let mut loaded_pots_with_balance = vec![];

        for (pot, users) in loaded_pots {
            let balance = pot_api_state
                .expense_service
                .get_pot_net_balance(pot.id(), subject_id)
                .await
                .map_err(check_error)?;

            loaded_pots_with_balance.push((pot, users, balance))
        }

        Ok((
            StatusCode::OK,
            Json(PotDTO::from_vec(loaded_pots_with_balance, all_currencies)),
        ))
    }

    /// Adds the given users to the pot if Bearer is the owner of that pot.
    #[utoipa::path(
            put,
            path = "/pots/{pot_id}",
            tag = "Pots",
            responses(
                (status = 200, description = "The users have successfully been added to the pot"),
                (status = 403, description = "The users could not be added due to the caller not being the owner of the given pot."),
                (status = 409, description = "The users were already added to the pot."),
                (status = 423, description = "The users can't be added, as the pot is archived."),
                (status = 500, description = "An internal server error occurred")
            ),
            request_body = Vec<AddUserToPotDTO>,
            params(
                ("pot_id" = i32, Path, description = "Pot database id for the pot.  ")
            ),
            security(
                    ("bearer" = [])
            )
        )]
    pub async fn add_users_to_pot(
        State(pot_api_state): State<Arc<PotApiState>>,
        Path(pot_id): Path<i32>,
        part: Parts,
        Json(add_user_to_pot_dto): Json<Vec<AddUserToPotDTO>>,
    ) -> Result<ApiResponse<String>, ApiResponse<String>> {
        let subject_id = get_sub_claim(&part)?;

        // create a vector of PotToUser so that they can be added to the database
        let mut new_user_ids = vec![];
        for user in add_user_to_pot_dto {
            let new_user_id = user.user_id();
            new_user_ids.push(PotToUser::new(pot_id, new_user_id));
        }

        let result = pot_api_state
            .pot_service
            .add_users_to_pot(new_user_ids, subject_id)
            .await
            .map_err(check_error)?;

        if !result {
            return Ok((
                StatusCode::FORBIDDEN,
                Json(format!("Users could not be added to pot {}", pot_id)),
            ));
        }

        Ok((
            StatusCode::OK,
            Json(format!("Users successfully added to pot {}", pot_id)),
        ))
    }

    #[utoipa::path(
            put,
            path = "/pots/{pot_id}/archive",
            tag = "Pots",
            responses(
                (status = 200, description = "The pot has been successfully archived."),
                (status = 403, description = "The pot could not be archived due to the caller not being the owner of the given pot."),
                (status = 409, description = "The pot was already archived."),
                (status = 500, description = "An internal server error occurred")
            ),
            params(
                ("pot_id" = i32, Path, description = "Pot database id for the pot")
            ),
            security(
                    ("bearer" = [])
            )
    )]
    pub async fn archive(
        State(pot_api_state): State<Arc<PotApiState>>,
        Path(pot_id): Path<i32>,
        part: Parts,
    ) -> Result<ApiResponse<String>, ApiResponse<String>> {
        let subject_id = get_sub_claim(&part)?;

        let result = pot_api_state
            .pot_service
            .archive(pot_id, subject_id)
            .await
            .map_err(check_error)?;

        if !result {
            return Ok((
                StatusCode::FORBIDDEN,
                Json(format!(
                    "Pot {} could not be archived by user.",
                    pot_id
                )),
            ));
        }

        Ok((
            StatusCode::OK,
            Json(format!(
                "Pot {} was successfully archived by user.",
                pot_id
            )),
        ))
    }

    #[utoipa::path(
        put,
        path = "/pots/{pot_id}/unarchive",
        tag = "Pots",
        responses(
                (status = 200, description = "The pot has been successfully unarchived."),
                (status = 403, description = "The pot could not be unarchived due to the caller not being the owner of the given pot."),
                (status = 409, description = "The pot is not archived."),
                (status = 500, description = "An internal server error occurred")
        ),
        params(
                ("pot_id" = i32, Path, description = "Pot database id for the pot")
        ),
        security(
                    ("bearer" = [])
        )
    )]
    pub async fn unarchive(
        State(pot_api_state): State<Arc<PotApiState>>,
        Path(pot_id): Path<i32>,
        part: Parts,
    ) -> Result<ApiResponse<String>, ApiResponse<String>> {
        let subject_id = get_sub_claim(&part)?;

        let result = pot_api_state
            .pot_service
            .unarchive(pot_id, subject_id)
            .await
            .map_err(check_error)?;

        if !result {
            return Ok((
                StatusCode::FORBIDDEN,
                Json(format!(
                    "Pot {} could not be unarchived by user.",
                    pot_id
                )),
            ));
        }

        Ok((
            StatusCode::OK,
            Json(format!(
                "Pot {} was successfully unarchived by user.",
                pot_id
            )),
        ))
    }

    /// Marks the given pot as paid. Only the pot owner can do this.
    /// This is a function used to mark the pot as paid with a single button click.
    /// No restore is possible as of now. So the pot is considered final.
    #[utoipa::path(
        put,
        path = "/pots/{pot_id}/pay",
        tag = "Pots",
        responses(
                (status = 200, description = "The pot has been successfully payed."),
                (status = 403, description = "The pot could not be payed."),
                (status = 500, description = "An internal server error occurred")
        ),
        params(
                ("pot_id" = i32, Path, description = "Pot database id for the pot")
        ),
        security(
                    ("bearer" = [])
        )
    )]
    pub async fn pay_pot(
        State(pot_api_state): State<Arc<PotApiState>>,
        Path(pot_id): Path<i32>,
        part: Parts
    ) -> Result<ApiResponse<String>, ApiResponse<String>> {
        let subject_id = get_sub_claim(&part)?;

        let result = pot_api_state.pot_service.pay_pot(pot_id, subject_id).await;

        if result.is_ok() {
            return Ok((StatusCode::OK, Json("Pot paid successfully".to_string())));
        }

        Err((StatusCode::FORBIDDEN, Json("Pot could not be paid".to_string())))
    }

    /// Removes the given user from the pot, if Bearer is the owner of that pot.
    #[utoipa::path(
            put,
            path = "/pots/{pot_id}/remove_user",
            tag = "Pots",
            responses(
                (status = 204, description = "The user has successfully been removed from the pot"),
                (status = 403, description = "The user could not be removed due to the caller not being the owner of the given pot."),
                (status = 404, description = "The user is not part of the pot."),
                (status = 423, description = "The user can't be removed, as the pot is archived."),
                (status = 500, description = "An internal server error occurred")
            ),
            request_body = RemoveUserFromPotDTO,
            params(
                ("pot_id" = i32, Path, description = "Pot database id for the pot.  ")
            ),
            security(
                    ("bearer" = [])
            )
    )]
    pub async fn remove_user_from_pot(
        State(pot_api_state): State<Arc<PotApiState>>,
        Path(pot_id): Path<i32>,
        part: Parts,
        Json(remove_user_from_pot_dto): Json<RemoveUserFromPotDTO>,
    ) -> Result<ApiResponse<String>, ApiResponse<String>> {
        let subject_id = get_sub_claim(&part)?;
        let new_user_id = remove_user_from_pot_dto.user_id();

        let result = pot_api_state
            .pot_service
            .remove_user_from_pot(PotToUser::new(pot_id, new_user_id), subject_id)
            .await
            .map_err(check_error)?;

        if !result {
            return Ok((
                StatusCode::FORBIDDEN,
                Json(format!(
                    "User {} could not be removed from pot {}",
                    new_user_id, pot_id
                )),
            ));
        }

        Ok((
            StatusCode::NO_CONTENT,
            Json(format!(
                "User {} successfully removed from pot {}",
                new_user_id, pot_id
            )),
        ))
    }

    /// Adds a new expense in the name of the user from the Bearer token to the pot with the given
    /// pot_id if it exists.
    #[utoipa::path(
        post,
        path = "/pots/{pot_id}",
        tag = "Pots",
        responses(
            (status = 201, description = "Indicates that the expense has been created for the given pot.",body = ExpenseDTO),
            (status = 404, description = "Indicates that the pot for this expense does not exist."),
            (status = 423, description = "The user can't be added, as the pot is archived."),
        ),
        request_body = NewExpenseDTO,
        params(
            ("pot_id" = i32, Path, description = "Pot database id for the pot.  ")
        ),
        security(
            ("bearer" = [])
        )
    )]
    pub async fn add_expense(
        State(pot_api_state): State<Arc<PotApiState>>,
        Path(pot_id): Path<i32>,
        part: Parts,
        Json(new_expense): Json<NewExpenseDTO>,
    ) -> Result<ApiResponse<ExpenseDTO>, ApiResponse<String>> {
        let subject_id = get_sub_claim(&part)?;

        let loaded_pot = pot_api_state
            .pot_service
            .get_pot_by_id(pot_id, subject_id)
            .await
            .map_err(check_error)?;

        if loaded_pot.is_archived() {
            return Err((StatusCode::LOCKED, Json("Pot is archived".to_string())));
        }

        let expense_splits_result = pot_api_state
            .expense_service
            .create_expense(
                new_expense.to_db(loaded_pot.id(), subject_id),
                new_expense.splits_to_new_db(),
            )
            .await
            .map_err(check_error)?;

        let expense = expense_splits_result.0;
        let splits = expense_splits_result.1;
        let currency = expense_splits_result.2;

        Ok((
            StatusCode::CREATED,
            Json(ExpenseDTO::from(expense, currency, splits, subject_id)),
        ))
    }

    /// Gets the sum of all expenses for the given user of the given pot.
    #[utoipa::path(
        get,
        path = "/pots/{pot_id}",
        tag = "Pots",
        responses(
            (
                status = 200,
                description = "The expenses for the pot with the given id",
                body = Vec<ExpenseDTO>
            ),
            (
                status = 404,
                description = "Indicates that the desired pot does not exists"
            )
        ),
        params(
            ("pot_id" = i32, Path, description = "Pot database id for the pot.  ")
        ),
        security(
            ("bearer" = [])
        )
    )]
    pub async fn get_pot_expenses(
        State(pot_api_service): State<Arc<PotApiState>>,
        Path(pot_id): Path<i32>,
        parts: Parts,
    ) -> Result<ApiResponse<Vec<ExpenseDTO>>, ApiResponse<String>> {
        let subject_id = get_sub_claim(&parts)?;

        let result = pot_api_service
            .expense_service
            .get_expenses_by_pot_id(pot_id, subject_id)
            .await
            .map_err(check_error)?;

        Ok((
            StatusCode::OK,
            Json(ExpenseDTO::from_vec(result, subject_id)),
        ))
    }

    /// Deletes the given pot if it does not contain any outstanding expenses and the caller
    /// is the pot's owner.
    #[utoipa::path(
        delete,
        path = "/pots/{pot_id}",
        tag = "Pots",
        responses(
            (status = 204, description = "The pot has been deleted."),
            (status = 403, description = "Indicates that the user is not authorized to delete the given pot."),
            (status = 404, description = "Indicates that the desired pot does not exists."),
            (status = 409, description = "Indicates that the desired pot can't be deleted.")
        ),
        params(
            ("pot_id" = i32, Path, description = "Pot database id for the pot.  ")
        ),
        security(
            ("bearer" = [])
        )
    )]
    pub async fn delete_pot(
        State(pot_api_service): State<Arc<PotApiState>>,
        Path(pot_id): Path<i32>,
        parts: Parts,
    ) -> Result<ApiResponse<String>, ApiResponse<String>> {
        let subject_id = get_sub_claim(&parts)?;

        pot_api_service
            .pot_service
            .try_delete_pot(pot_id, subject_id)
            .await
            .map_err(check_error)?;

        Ok((
            StatusCode::NO_CONTENT,
            Json(format!("Deleted pot {}", pot_id)),
        ))
    }
}