# openapi/expense_tracker_openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "expense_tracker",
    "description": "A REST Api that offers various endpoints for handling shared expenses.",
    "contact": {
      "name": "Ciel1996"
    },
    "license": {
      "name": "GPL-3.0-or-later",
      "identifier": "GPL-3.0-or-later"
    },
    "version": "1.2.0"
  },
  "paths": {
    "/api/v1/currencies": {
      "get": {
        "tags": [
          "Currency"
        ],
        "summary": "Gets a list of all currently known currencies.",
        "operationId": "get_currencies",
        "responses": {
          "200": {
            "description": "All currencies known to the system",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CurrencyDTO"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "Currency"
        ],
        "summary": "Creates a new currency, if symbol is not yet in use.",
        "operationId": "create_currency",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NewCurrencyDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "The currency has been created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NewCurrencyDTO"
                }
              }
            }
          },
          "409": {
            "description": "Detected a conflict, as the symbol is already known."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/api/v1/pots": {
      "get": {
        "tags": [
          "Pots"
        ],
        "summary": "Gets the list of all pots the bearer can view.",
        "operationId": "get_pots",
        "responses": {
          "200": {
            "description": "The list of known pots.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PotDTO"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "Pots"
        ],
        "summary": "Creates a pot from the given DTO for the bearer.",
        "operationId": "create_pot",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NewPotDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "The pot has been created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PotDTO"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/pots/{pot_id}": {
      "get": {
        "tags": [
          "Pots"
        ],
        "summary": "Gets the sum of all expenses for the given user of the given pot.",
        "operationId": "get_pot_expenses",
        "parameters": [
          {
            "name": "pot_id",
            "in": "path",
            "description": "Pot database id for the pot.  ",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The expenses for the pot with the given id",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ExpenseDTO"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Indicates that the desired pot does not exists"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "put": {
        "tags": [
          "Pots"
        ],
        "summary": "Adds the given users to the pot if Bearer is the owner of that pot.",
        "operationId": "add_users_to_pot",
        "parameters": [
          {
            "name": "pot_id",
            "in": "path",
            "description": "Pot database id for the pot.  ",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/AddUserToPotDTO"
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "The users have successfully been added to the pot"
          },
          "403": {
            "description": "The users could not be added due to the caller not being the owner of the given pot."
          },
          "409": {
            "description": "The users were already added to the pot."
          },
          "423": {
            "description": "The users can't be added, as the pot is archived."
          },
          "500": {
            "description": "An internal server error occurred"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "Pots"
        ],
        "summary": "Adds a new expense in the name of the user from the Bearer token to the pot with the given\npot_id if it exists.",
        "operationId": "add_expense",
        "parameters": [
          {
            "name": "pot_id",
            "in": "path",
            "description": "Pot database id for the pot.  ",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NewExpenseDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Indicates that the expense has been created for the given pot.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExpenseDTO"
                }
              }
            }
          },
          "404": {
            "description": "Indicates that the pot for this expense does not exist."
          },
          "423": {
            "description": "The user can't be added, as the pot is archived."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "tags": [
          "Pots"
        ],
        "summary": "Deletes the given pot if it does not contain any outstanding expenses and the caller\nis the pot's owner.",
        "operationId": "delete_pot",
        "parameters": [
          {
            "name": "pot_id",
            "in": "path",
            "description": "Pot database id for the pot.  ",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "The pot has been deleted."
          },
          "403": {
            "description": "Indicates that the user is not authorized to delete the given pot."
          },
          "404": {
            "description": "Indicates that the desired pot does not exists."
          },
          "409": {
            "description": "Indicates that the desired pot can't be deleted."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/pots/{pot_id}/archive": {
      "put": {
        "tags": [
          "Pots"
        ],
        "operationId": "archive",
        "parameters": [
          {
            "name": "pot_id",
            "in": "path",
            "description": "Pot database id for the pot",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The pot has been successfully archived."
          },
          "403": {
            "description": "The pot could not be archived due to the caller not being the owner of the given pot."
          },
          "409": {
            "description": "The pot was already archived."
          },
          "500": {
            "description": "An internal server error occurred"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/pots/{pot_id}/pay": {
      "put": {
        "tags": [
          "Pots"
        ],
        "summary": "Marks the given pot as paid. Only the pot owner can do this.\nThis is a function used to mark the pot as paid with a single button click.\nNo restore is possible as of now. So the pot is considered final.",
        "operationId": "pay_pot",
        "parameters": [
          {
            "name": "pot_id",
            "in": "path",
            "description": "Pot database id for the pot",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The pot has been successfully payed."
          },
          "403": {
            "description": "The pot could not be payed."
          },
          "500": {
            "description": "An internal server error occurred"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/pots/{pot_id}/remove_user": {
      "put": {
        "tags": [
          "Pots"
        ],
        "summary": "Removes the given user from the pot, if Bearer is the owner of that pot.",
        "operationId": "remove_user_from_pot",
        "parameters": [
          {
            "name": "pot_id",
            "in": "path",
            "description": "Pot database id for the pot.  ",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RemoveUserFromPotDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "The user has successfully been removed from the pot"
          },
          "403": {
            "description": "The user could not be removed due to the caller not being the owner of the given pot."
          },
          "404": {
            "description": "The user is not part of the pot."
          },
          "423": {
            "description": "The user can't be removed, as the pot is archived."
          },
          "500": {
            "description": "An internal server error occurred"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/pots/{pot_id}/unarchive": {
      "put": {
        "tags": [
          "Pots"
        ],
        "operationId": "unarchive",
        "parameters": [
          {
            "name": "pot_id",
            "in": "path",
            "description": "Pot database id for the pot",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The pot has been successfully unarchived."
          },
          "403": {
            "description": "The pot could not be unarchived due to the caller not being the owner of the given pot."
          },
          "409": {
            "description": "The pot is not archived."
          },
          "500": {
            "description": "An internal server error occurred"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "AddUserToPotDTO": {
        "type": "object",
        "required": [
          "user_id"
        ],
        "properties": {
          "user_id": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
      "CurrencyDTO": {
        "type": "object",
        "description": "DTO representing a currency.",
        "required": [
          "id",
          "name",
          "symbol"
        ],
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string"
          },
          "symbol": {
            "type": "string"
          }
        }
      },
...
      "NewCurrencyDTO": {
        "type": "object",
        "description": "DTO representing a new currency.",
        "required": [
          "name",
          "symbol"
        ],
        "properties": {
          "name": {
            "type": "string"
          },
          "symbol": {
            "type": "string"
          }
        }
      },
...
      "NewPotDTO": {
        "type": "object",
        "description": "DTO used when creating a new Pot.",
        "required": [
          "name",
          "default_currency_id"
        ],
        "properties": {
          "default_currency_id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "maxLength": 24
          }
        }
      },
...
      "RemoveUserFromPotDTO": {
        "type": "object",
        "required": [
          "user_id"
        ],
        "properties": {
          "user_id": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}